Update notes:
I updated the 3DComic Blender toolkit for blender 2.91 to support the new load screen and panel bonuses I'm doing on 3Dcomic.shop/inkbots.   There is also a new folder structure that gets automatically created when you make a new 3D comic.  
This is to prepare for using shared assets between panels in both the online 3D Comic page and in the 3D Comic .blend file using blender linked libraries with armature overrides 
(I'll share it when I'm happy with it)